Move VM const pool decoder into core
