+++ docs/internal/testing/checklists/architecture-improvements.md
@@ -63,6 +63,7 @@ Issue #51 init benchmark baseline evidence (2026-04-28):
 - [x] `ARCH-RTCORE-27` Move VM frame-stack and local-frame state into `trust-runtime-core`, preserve `trust-runtime::runtime::vm::frames` as a compatibility re-export, and validate local-reference bounds/load/store behavior, frame-stack depth enforcement, `bytecode_vm_core`, runtime VM lib filters, clippy, full-map doctor, and diagram drift gates.
 - [x] `ARCH-RTCORE-28` Move pure VM dispatch-op helpers into `trust-runtime-core`, narrow the binary-op helper from full `Runtime` access to `DateTimeProfile`, preserve `trust-runtime::runtime::vm::dispatch_ops` as a compatibility re-export, and validate core dispatch-op tests plus bytecode VM and runtime VM filters.
 - [x] `ARCH-RTCORE-29` Move bytecode type-table `SIZEOF` evaluation into `trust-runtime-core` without adding a HIR dependency to the pure core path, keep the HIR-gated `SizeOfError` mapper in the runtime compatibility module, and validate focused core/runtime `SIZEOF` gates plus clippy, full-map doctor, and diagram drift.
+- [x] `ARCH-RTCORE-30` Move VM const-pool decoding into `trust-runtime-core`, preserve `trust-runtime::runtime::vm::const_pool` as a compatibility re-export, replace `std` dependencies with `core`/`alloc`, and validate primitive/enum/alias/subrange const decoding plus focused bytecode VM const-error gates.
 - [x] `ARCH-FULL-01` Add `full-architecture-refactor-program-checklist.md` as the umbrella board that states the runtime-core split is one branch of the full SOLID/KISS and zero-silent-bug program, not the whole program.
 - [x] `ARCH-FULL-02` Add `architecture-doctor-full-map-execution-checklist.md` so `architecture-doctor --full-map` becomes a concrete prerequisite for enforcing workspace, runtime-core, CLI, host-surface, KISS, API, and diagram rules.
 - [x] `ARCH-FULL-03` Add `hir-mutation-hardening-execution-checklist.md` for the F2 silent-bug gap: `symbol_import`, `type_check::const_eval`, and aggregate initializer validation mutation survivors.
