+++ docs/internal/testing/checklists/runtime-core-host-split-execution-checklist.md
@@ -270,7 +270,7 @@ Move lowest-risk portable pieces first.
 ## Phase 5 - Move VM Execution Core
 
 - [x] `RTSPLIT-P5-001` Identify VM modules that are pure execution versus host assembly/lowering. Evidence: pure leaf modules ready for core are VM trap/error mapping and operand stack; pure candidates needing additional seams are dispatch ops, SIZEOF helpers, const-pool decode, and frame stack pieces; runtime-coupled modules are dispatch, dispatch refs, call, local init, and register IR execution because they still take `Runtime`, debug hooks, or host storage/initializer services; lowering/profile/tier1/test-fixture modules stay split by their existing ownership boundaries.
-- [ ] `RTSPLIT-P5-002` Move VM dispatch/execution pieces that do not require host services. Progress: VM trap/error mapping, operand stack, opcode-length helper, borrowed-value materialization helper, frame-stack/local-frame state, pure dispatch arithmetic/jump/operand-read helpers, and bytecode type-table `SIZEOF` evaluation moved to `trust-runtime-core`; larger dispatch, call, local init, and register IR execution remain host-side until their `Runtime` and debug/initializer dependencies are split.
+- [ ] `RTSPLIT-P5-002` Move VM dispatch/execution pieces that do not require host services. Progress: VM trap/error mapping, operand stack, opcode-length helper, borrowed-value materialization helper, frame-stack/local-frame state, pure dispatch arithmetic/jump/operand-read helpers, bytecode type-table `SIZEOF` evaluation, and const-pool decoding moved to `trust-runtime-core`; larger dispatch, call, local init, and register IR execution remain host-side until their `Runtime` and debug/initializer dependencies are split.
 - [ ] `RTSPLIT-P5-003` Keep host assembly, IO driver invocation, CLI project loading, and runtime config loading in `trust-runtime`.
 - [ ] `RTSPLIT-P5-004` Introduce service ports only where the VM truly needs host callbacks.
 - [ ] `RTSPLIT-P5-005` Do not let VM code import `web`, `control`, `debug`, `runtime_cloud`, or host IO implementations through the new core.
@@ -292,6 +292,8 @@ Move lowest-risk portable pieces first.
 - [x] `RTSPLIT-P5-VM-LEAF-009` Verify the VM dispatch-op movement slice. Evidence: `just fmt`, `cargo test -p trust-runtime-core vm -- --nocapture`, `cargo check -p trust-runtime-core --no-default-features`, `cargo check -p trust-runtime --lib`, `cargo test -p trust-runtime --test bytecode_vm_core -- --nocapture`, `cargo test -p trust-runtime runtime::vm:: --lib -- --nocapture`, `cargo clippy -p trust-runtime-core -p trust-runtime --lib -- -D warnings`, `cargo run -p xtask -- architecture-doctor --full-map`, `scripts/render_diagrams.sh`, and `python scripts/check_diagram_drift.py` pass.
 - [x] `RTSPLIT-P5-VM-LEAF-010` Move bytecode type-table `SIZEOF` evaluation into `trust-runtime-core` without importing HIR. Evidence: `crates/trust-runtime-core/src/vm/dispatch_sizeof.rs` now owns `sizeof_type_from_table` and uses a local `core::mem::size_of::<usize>()` reference-handle constant matching the HIR constant; `trust-runtime/src/runtime/vm/dispatch_sizeof.rs` keeps only the `SizeOfError` mapper because that value-size API remains HIR-feature-gated.
 - [x] `RTSPLIT-P5-VM-LEAF-011` Verify the VM `SIZEOF` movement slice. Evidence: `just fmt`, `cargo test -p trust-runtime-core vm -- --nocapture`, `cargo check -p trust-runtime-core --no-default-features`, `cargo check -p trust-runtime --lib`, `cargo test -p trust-runtime --test bytecode_vm_core sizeof -- --nocapture`, `cargo clippy -p trust-runtime-core -p trust-runtime --lib -- -D warnings`, `cargo run -p xtask -- architecture-doctor --full-map`, `scripts/render_diagrams.sh`, and `python scripts/check_diagram_drift.py` pass.
+- [x] `RTSPLIT-P5-VM-LEAF-012` Move VM const-pool decoding into `trust-runtime-core`. Evidence: `crates/trust-runtime-core/src/vm/const_pool.rs` now owns `decode_const_pool_entries`, primitive constant decoding, enum constant decoding, alias/subrange const type resolution, and const-payload validation using `core`/`alloc` instead of `std`; `trust-runtime/src/runtime/vm/const_pool.rs` is a compatibility re-export.
+- [x] `RTSPLIT-P5-VM-LEAF-013` Verify the VM const-pool movement slice. Evidence: `just fmt`, `cargo test -p trust-runtime-core vm -- --nocapture`, `cargo check -p trust-runtime-core --no-default-features`, `cargo check -p trust-runtime --lib`, `cargo test -p trust-runtime --test bytecode_vm_core const -- --nocapture`, `cargo clippy -p trust-runtime-core -p trust-runtime --lib -- -D warnings`, `cargo run -p xtask -- architecture-doctor --full-map`, `scripts/render_diagrams.sh`, and `python scripts/check_diagram_drift.py` pass.
 
 ### Phase 5 Exit Gate
